Move `CodableFeedStoreClone` to its own file in production
